feat: move sv back into public
